+++ pybamm/expression_tree/scalar.py
@@ -46,6 +46,9 @@ class Scalar(pybamm.Symbol):
 
         super().__init__(name, domain=domain, units=units)
 
+    def __str__(self):
+        return str(self.value)
+
     @property
     def value(self):
         """the value returned by the node when evaluated"""