+++ pybamm/models/full_battery_models/base_battery_model.py
@@ -6,7 +6,7 @@ import pybamm
 import warnings
 
 
-class Options(pybamm.FuzzyDict):
+class BatteryModelOptions(pybamm.FuzzyDict):
     """
     Attributes
     ----------
@@ -16,6 +16,9 @@ class Options(pybamm.FuzzyDict):
         be set are listed below. Note that not all of the options are compatible with
         each other and with all of the models implemented in PyBaMM. Each option is
         optional and takes a default value if not provided.
+        In general, the option provided must be a string, but there are some cases
+        where a 2-tuple of strings can be provided instead to indicate a different
+        option for the negative and positive electrodes.
 
             * "cell geometry" : str
                 Sets the geometry of the cell. Can be "pouch" (default) or
@@ -41,16 +44,18 @@ class Options(pybamm.FuzzyDict):
                 variable for instead of solving in PyBaMM. The entries of the lists
                 are strings that correspond to the submodel names in the keys
                 of `self.submodels`.
-            * "interfacial surface area" : str
-                Sets the model for the interfacial surface area. Can be "constant"
-                (default) or "varying". Not currently implemented in any of the models.
+            * "hydrolysis" : str
+                Whether to include hydrolysis in the model. Only implemented for
+                lead-acid models. Can be "false" (default) or "true". If "true", then
+                "surface form" cannot be 'false'.
             * "lithium plating" : str, optional
                 Sets the model for lithium plating. Can be "none" (default),
                 "reversible" or "irreversible".
             * "loss of active material" : str
                 Sets the model for loss of active material. Can be "none" (default),
-                "positive", "negative" or "both" to enable it for the specific
-                electrode.
+                "stress-driven", or "reaction-driven".
+                A 2-tuple can be provided for different behaviour in negative and
+                positive electrodes.
             * "operating mode" : str
                 Sets the operating mode for the model. Can be "current" (default),
                 "voltage" or "power". Alternatively, the operating mode can be
@@ -68,15 +73,12 @@ class Options(pybamm.FuzzyDict):
                 (default), "user" or "no particles". For the "user" option the surface
                 area per unit volume can be passed as a parameter, and is therefore not
                 necessarily consistent with the particle shape.
-            * "particle cracking" : str
-                Sets the model to account for mechanical effects and particle
-                cracking. Can be "none", "no cracking", "negative", "positive" or
-                "both".
-                All options other than "none" account for the effects of swelling
-                of electrode particles, cell thickness change, and stress-assisted
-                diffusion. The options "negative", "positive" or "both" additionally
-                account for crack propagation in the negative, positive or both
-                electrodes, respectively.
+            * "particle mechanics" : str
+                Sets the model to account for mechanical effects such as particle
+                swelling and cracking. Can be "none" (default), "swelling only",
+                or "swelling and cracking".
+                A 2-tuple can be provided for different behaviour in negative and
+                positive electrodes.
             * "SEI" : str
                 Set the SEI submodel to be used. Options are:
 
@@ -119,10 +121,6 @@ class Options(pybamm.FuzzyDict):
             * "SEI porosity change" : str
                 Whether to include porosity change due to SEI formation, can be "false"
                 (default) or "true".
-            * "side reactions" : list
-                Contains a list of any side reactions to include. Default is []. If this
-                list is not empty (i.e. side reactions are included in the model), then
-                "surface form" cannot be 'false'.
             * "surface form" : str
                 Whether to use the surface formulation of the problem. Can be "false"
                 (default), "differential" or "algebraic".
@@ -140,7 +138,7 @@ class Options(pybamm.FuzzyDict):
 
     def __init__(self, extra_options):
         self.possible_options = {
-            "surface form": ["false", "differential", "algebraic"],
+            "cell geometry": ["arbitrary", "pouch"],
             "convection": ["none", "uniform transverse", "full transverse"],
             "current collector": [
                 "uniform",
@@ -148,9 +146,27 @@ class Options(pybamm.FuzzyDict):
                 "potential pair quite conductive",
             ],
             "dimensionality": [0, 1, 2],
-            "interfacial surface area": ["constant", "varying"],
-            "thermal": ["isothermal", "lumped", "x-lumped", "x-full"],
-            "cell geometry": ["arbitrary", "pouch"],
+            "electrolyte conductivity": [
+                "default",
+                "full",
+                "leading order",
+                "composite",
+                "integrated",
+            ],
+            "hydrolysis": ["true", "false"],
+            "lithium plating": ["none", "reversible", "irreversible"],
+            "lithium plating porosity change": ["true", "false"],
+            "loss of active material": ["none", "stress-driven", "reaction-driven"],
+            "operating mode": ["current", "voltage", "power"],
+            "particle": [
+                "Fickian diffusion",
+                "fast diffusion",
+                "uniform profile",
+                "quadratic profile",
+                "quartic profile",
+            ],
+            "particle mechanics": ["none", "swelling only", "swelling and cracking"],
+            "particle shape": ["spherical", "user", "no particles"],
             "SEI": [
                 "none",
                 "constant",
@@ -162,63 +178,38 @@ class Options(pybamm.FuzzyDict):
             ],
             "SEI film resistance": ["none", "distributed", "average"],
             "SEI porosity change": ["true", "false"],
-            "lithium plating": ["none", "reversible", "irreversible"],
-            "loss of active material": ["none", "negative", "positive", "both"],
-            "operating mode": ["current", "voltage", "power"],
-            "particle cracking": [
-                "none",
-                "no cracking",
-                "negative",
-                "positive",
-                "both",
-            ],
-            "lithium plating porosity change": ["true", "false"],
-            "particle": [
-                "Fickian diffusion",
-                "fast diffusion",
-                "uniform profile",
-                "quadratic profile",
-                "quartic profile",
-            ],
-            "particle shape": ["spherical", "user", "no particles"],
-            "electrolyte conductivity": [
-                "default",
-                "full",
-                "leading order",
-                "composite",
-                "integrated",
-            ],
+            "surface form": ["false", "differential", "algebraic"],
+            "thermal": ["isothermal", "lumped", "x-lumped", "x-full"],
             "total interfacial current density as a state": ["true", "false"],
         }
 
         default_options = {
-            "operating mode": "current",
-            "dimensionality": 0,
-            "surface form": "false",
-            "convection": "none",
-            "side reactions": [],
-            "interfacial surface area": "constant",
-            "current collector": "uniform",
-            "particle": "Fickian diffusion",
-            "particle shape": "spherical",
-            "electrolyte conductivity": "default",
-            "thermal": "isothermal",
             "cell geometry": "none",
+            "convection": "none",
+            "current collector": "uniform",
+            "dimensionality": 0,
+            "electrolyte conductivity": "default",
             "external submodels": [],
-            "SEI": "none",
+            "hydrolysis": "false",
             "lithium plating": "none",
-            "SEI porosity change": "false",
             "lithium plating porosity change": "false",
             "loss of active material": "none",
-            "working electrode": "none",
-            "particle cracking": "none",
+            "operating mode": "current",
+            "particle": "Fickian diffusion",
+            "particle mechanics": "none",
+            "particle shape": "spherical",
+            "SEI": "none",
+            "SEI porosity change": "false",
+            "surface form": "false",
+            "thermal": "isothermal",
             "total interfacial current density as a state": "false",
+            "working electrode": "none",
         }
 
         # Change the default for cell geometry based on which thermal option is provided
         extra_options = extra_options or {}
-        thermal_option = extra_options.get("thermal", "none")
         # return "none" if option not given
+        thermal_option = extra_options.get("thermal", "none")
         if thermal_option in ["none", "isothermal", "lumped"]:
             default_options["cell geometry"] = "arbitrary"
         else:
@@ -228,13 +219,23 @@ class Options(pybamm.FuzzyDict):
 
         # Change the default for SEI film resistance based on which SEI option is
         # provided
-        # extra_options = extra_options or {}
-        sei_option = extra_options.get("SEI", "none")
         # return "none" if option not given
+        sei_option = extra_options.get("SEI", "none")
         if sei_option == "none":
             default_options["SEI film resistance"] = "none"
         else:
             default_options["SEI film resistance"] = "distributed"
+        # The "SEI film resistance" option will still be overridden by extra_options if
+        # provided
+
+        # Change the default for swelling based on which LAM option is
+        # provided
+        # return "none" if option not given
+        lam_option = extra_options.get("loss of active material", "none")
+        if "stress-driven" in lam_option:
+            default_options["particle mechanics"] = "swelling only"
+        else:
+            default_options["particle mechanics"] = "none"
         # The "SEI film resistance" option will still be overridden by extra_options if
         # provided
 
@@ -244,11 +245,17 @@ class Options(pybamm.FuzzyDict):
             if name in default_options:
                 options[name] = opt
             else:
-                raise pybamm.OptionError(
-                    "Option '{}' not recognised. Best matches are {}".format(
-                        name, options.get_best_matches(name)
+                if name == "particle cracking":
+                    raise pybamm.OptionError(
+                        "The 'particle cracking' option has been renamed. "
+                        "Use 'particle mechanics' instead."
                     )
-                )
+                else:
+                    raise pybamm.OptionError(
+                        "Option '{}' not recognised. Best matches are {}".format(
+                            name, options.get_best_matches(name)
+                        )
+                    )
 
         # If "SEI film resistance" is "distributed" then "total interfacial current
         # density as a state" must be "true"
@@ -294,18 +301,41 @@ class Options(pybamm.FuzzyDict):
             )
 
         for option, value in options.items():
-            if (
-                option == "side reactions"
-                or option == "external submodels"
-                or option == "working electrode"
-            ):
+            if option == "external submodels" or option == "working electrode":
                 pass
-            elif value not in self.possible_options[option]:
-                if not (option == "operating mode" and callable(value)):
-                    raise pybamm.OptionError(
-                        f"\n'{value}' is not recognized in option '{option}'. "
-                        f"Possible values are {self.possible_options[option]}"
-                    )
+            else:
+                if isinstance(value, str) or option in [
+                    "dimensionality",
+                    "operating mode",
+                ]:  # some options don't take strings
+                    value = (value,)
+                else:
+                    if not (
+                        (
+                            option
+                            in [
+                                "loss of active material",
+                                "particle mechanics",
+                                "particle",
+                            ]
+                            and isinstance(value, tuple)
+                            and len(value) == 2
+                        )
+                    ):
+                        # more possible options that can take 2-tuples to be added
+                        # as they come
+                        raise pybamm.OptionError(
+                            f"\n'{value}' is not recognized in option '{option}'. "
+                            "Values must be strings or (in some cases) "
+                            "2-tuples of strings"
+                        )
+                for val in value:
+                    if val not in self.possible_options[option]:
+                        if not (option == "operating mode" and callable(val)):
+                            raise pybamm.OptionError(
+                                f"\n'{val}' is not recognized in option '{option}'. "
+                                f"Possible values are {self.possible_options[option]}"
+                            )
 
         super().__init__(options.items())
 
@@ -413,7 +443,7 @@ class BaseBatteryModel(pybamm.BaseModel):
 
     @options.setter
     def options(self, extra_options):
-        options = Options(extra_options)
+        options = BatteryModelOptions(extra_options)
 
         # Options that are incompatible with models
         if isinstance(self, pybamm.lithium_ion.BaseModel):
@@ -442,14 +472,14 @@ class BaseBatteryModel(pybamm.BaseModel):
         if (
             isinstance(self, (pybamm.lead_acid.LOQS, pybamm.lead_acid.Composite))
             and options["surface form"] == "false"
+            and options["hydrolysis"] == "true"
         ):
-            if len(options["side reactions"]) > 0:
-                raise pybamm.OptionError(
-                    """must use surface formulation to solve {!s} with side reactions
+            raise pybamm.OptionError(
+                """must use surface formulation to solve {!s} with hydrolysis
                     """.format(
-                        self
-                    )
+                    self
                 )
+            )
 
         self._options = options
 
@@ -640,6 +670,9 @@ class BaseBatteryModel(pybamm.BaseModel):
 
         pybamm.logger.debug("Setting SoC variables ({})".format(self.name))
         self.set_soc_variables()
+
+        pybamm.logger.debug("Setting degradation variables ({})".format(self.name))
+        self.set_degradation_variables()
 
         # Massive hack for consistent delta_phi = phi_s - phi_e with SPMe
         # This needs to be corrected
@@ -927,7 +960,7 @@ class BaseBatteryModel(pybamm.BaseModel):
         # mode)
         # A tolerance of 1 is sufficiently small since the dimensionless voltage is
         # scaled with the thermal voltage (0.025V) and hence has a range of around 60
-        tol = 1
+        tol = 5
         self.events.append(
             pybamm.Event(
                 "Minimum voltage switch",
@@ -946,6 +979,13 @@ class BaseBatteryModel(pybamm.BaseModel):
         # Power
         I_dim = self.variables["Current [A]"]
         self.variables.update({"Terminal power [W]": I_dim * V_dim})
+
+    def set_degradation_variables(self):
+        """
+        Set variables that quantify degradation.
+        This function is overriden by the base battery models
+        """
+        pass
 
     def set_soc_variables(self):
         """